Add HeaderAdmin, FooterAdmin components and update package dependencies
This commit introduces HeaderAdmin and FooterAdmin components to the layout package. It also changes the button in Submission1 and Submission2 components to ButtonUserForm. Package dependencies in package.json and package-lock.json files are updated - the notable addition is "@material-tailwind/react". The tailwind.config.js file is also changed to include the "@material-tailwind/react" package functionality.