Update loading state for country and variant data
Changed the initial state values for `countryData` and `variantData` from arrays to undefined. Added a check within a new useEffect Hook to set the `loading` state based on presence or absence of country and variant data. Modified the condition for displaying the form based on the `loading` state instead of data array lengths.