Implement submission step 1 and introduce cookie support
Added the functionality to handle and make API requests for the first step of the submission process, and implemented support for client cookies. Appropriate changes were made in the submission1.jsx component to send data to the backend when the "Next" button is clicked. We also used CookiesProvider in our root layout so that cookies can be used throughout our application.